Update after meeting on Apr 16th
During the topic leads meetign we decided to post the variable
availability to a website. Here the plots go to an md file.

Also, the Consistency study will be using the same variables as the
ClimateServices on, so their request is collapsed into one. However the
data request from the ClimateServices has been split into two
priorities, to track storage cost separately.

The storage_estimate and variable_availabilty scripts have been moved to
the code/ folder (they should still be executed from the root folder,
though)